automake foreign, recover INSTALL file, no NEWS is better
Add automake option foreign to configure.ac.  This tells the autotools
to not fail if a NEWS file is not found, so the empty NEWS file is no
longer created/installed.  And to not generate the boilerplate
standard GNU INSTALL file, which is hereby reverted to the old one
deleted while updating autotools: git checkout b6de1ecc -- INSTALL