dewarn fake key/button debug message
Debug messages in FakeThingsUp() attempt to print keysyms and button
numbers using printf format specifiers for integers (%x and %d). But the
type KeySym is likely to be unsigned long, which on many modern
platforms is not the same size as int. To eliminate both the compiler
warning and actual incorrect output, use %lx and %ld instead, along with
an explicit case to (unsigned) long to avoid issues if KeySym is not a
long.