Tell the window manager we are a dock
Set the trigger window's _NET_WM_WINDOW_TYPE property to
_NET_WM_WINDOW_TYPE_DOCK, which tells modern EWMH-compliant window
managers to give it special treatment. In particular, such windows are
stacked above all normal windows. Without this, the trigger window tends
to be obscured by maximized windows and those that overlap the display
edge. Relying on VisibilityNotify events to trigger re-raising the
window does not help, partly because the X server is not guaranteed to
generate them for InputOnly windows and partly because some compositing
window managers seem not to generate them at all.