Fixed issue https://github.com/dottedmag/x2x/issues/24
x2x won't move mouse to secondary screen if a primary window is too close
to the edge, on Ubuntu 18.04. This version added two options to address
this issue. -win-output makes the trigger window an InputOutput window
so visibility events are generated for the trigger window, and -win-transparent
makes the trigger window transparent, so not to bother users.