Set boundaries to honor -completeregion
If the -completeregion* options are used, set the location we warp to on
connect to be inside the specified region instead of at the edge of the
screen. Note that this considers only the edge to which we are warping.
A more complete fix would consider all four edges and move the pointer
in both directions, if necessary.

Additionally, return (or switch screens) as soon as the pointer reaches
the edge of the region, instead of waiting until it reaches the physical
edge of the from screen.